completed orders section

+++ src/store/reduxStore.js
@@ -5,6 +5,16 @@ const defaultState = {
   isLoggedIn: "",
 };
 const storeReducer = (state = defaultState, action) => {
+  if (action.type === "login") {
+    const data = action.data;
+    localStorage.setItem("userData", JSON.stringify(data));
+    return {
+      ...state,
+      userId: data.userId,
+      userEmail: data.userEmail,
+      isLoggedIn: true,
+    };
+  }
   if (action.type === "logout") {
     localStorage.clear();
     return { ...defaultState };
